refactor: 커뮤니티 서비스 패키지 정리와 아키텍처 문서 추가
