+++ src/main/java/com/raota/presentation/api/community/CommunityController.java
@@ -8,8 +8,8 @@ import com.raota.presentation.api.community.response.CommunityPostCardResponse;
 import com.raota.presentation.api.community.response.CommunityPostDetailResponse;
 import com.raota.presentation.api.community.response.CommunityRamenShopOptionResponse;
 import com.raota.domain.community.repository.query.PostQueryRepository;
-import com.raota.application.community.PostLikeService;
-import com.raota.application.community.PostService;
+import com.raota.application.community.service.PostLikeService;
+import com.raota.application.community.service.PostService;
 import com.raota.infrastructure.auth.LoginMember;
 import com.raota.presentation.common.ApiResponse;
 import com.raota.presentation.common.PageResponse;
