+++ src/test/java/com/raota/application/community/PostServiceTest.java
@@ -8,13 +8,12 @@ import com.raota.domain.community.model.PostCategory;
 import com.raota.presentation.api.community.request.CommunityPostCreateRequest;
 import com.raota.domain.community.repository.command.PostRepository;
 import com.raota.domain.community.repository.command.entity.PostEntity;
-import com.raota.application.community.PostService;
+import com.raota.application.community.service.PostService;
 import com.raota.domain.member.model.MemberProfile;
 import com.raota.domain.member.repository.MemberRepository;
 import com.raota.domain.ramenShop.repository.RamenShopRepository;
 import com.raota.domain.retrieval.event.PostIndexingAction;
 import com.raota.domain.retrieval.event.PostIndexingEvent;
-import java.time.LocalDateTime;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
